Added RotatableTowerLogic
Also removed the rotation logic from `TowerLogic`, so that only rotatable towers have that logic.
