+++ Assets/Prefabs/Towers/WheelTurret/WheelTurret.prefab
@@ -121,7 +121,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2462899330894574527}
-  - component: {fileID: 4634772014400295570}
+  - component: {fileID: 5783301673282278111}
   - component: {fileID: 3515612305777872196}
   - component: {fileID: 1759242131}
   - component: {fileID: 3531771237580499680}
@@ -148,7 +148,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &4634772014400295570
+--- !u!114 &5783301673282278111
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -157,7 +157,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 2983117196373584133}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 84a624fc6506f6f45a6bfd0335d144e7, type: 3}
+  m_Script: {fileID: 11500000, guid: 40bdf2e92d29c3a4e8319c1f05e5334f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   canInteract: 1
@@ -166,9 +166,7 @@ MonoBehaviour:
   bulletSpawnPoint: {fileID: 7245097492267856000}
   bulletSpeed: 5
   bulletDamage: 5
-  towerRotationPoint: {fileID: 3011992938053616436}
   rotAxis: {fileID: 769070955576516685}
-  rotationMaxSpeed: 0
   arrowPointer: {fileID: 8557351424284430801}
 --- !u!95 &3515612305777872196
 Animator:
