+++ Assets/Prefabs/Towers/tower_pendulum.prefab
@@ -41,7 +41,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 3686963311025437781}
-  - component: {fileID: 8782035920975212450}
+  - component: {fileID: 1350360576178508834}
   - component: {fileID: 8466527481623126138}
   - component: {fileID: 8528154779301624128}
   - component: {fileID: 154197252}
@@ -67,7 +67,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &8782035920975212450
+--- !u!114 &1350360576178508834
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -76,7 +76,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 423603099658726498}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 84a624fc6506f6f45a6bfd0335d144e7, type: 3}
+  m_Script: {fileID: 11500000, guid: 40bdf2e92d29c3a4e8319c1f05e5334f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   canInteract: 1
@@ -85,9 +85,7 @@ MonoBehaviour:
   bulletSpawnPoint: {fileID: 0}
   bulletSpeed: 0
   bulletDamage: 0
-  towerRotationPoint: {fileID: 6337973020223220929}
   rotAxis: {fileID: 0}
-  rotationMaxSpeed: 0
   arrowPointer: {fileID: 4911675975082342184}
 --- !u!136 &8466527481623126138
 CapsuleCollider:
